Build/Test Tools: Change the default value of `LOCAL_PHP`.
This changes the default value for `LOCAL_PHP` in the 5.8 branch from `latest` to `8.0-fpm` to reflect the highest version of PHP this branch will support (with beta support).

After this change, future updates to the `latest` container built and published by the `wpdev-docker-images` repository will not cause failures in this branch.

Follow up to [57198], [57198].

See #60095.

git-svn-id: https://develop.svn.wordpress.org/branches/5.8@57200 602fd350-edb4-49c9-b593-d223f7449a82