Build/Test Tools: Change the default value of `LOCAL_PHP`.
This changes the default value for `LOCAL_PHP` in the 5.9 branch from `latest` to `8.1-fpm` to reflect the highest version of PHP this branch will support (with beta support).

After this change, future updates to the `latest` container built and published by the `wpdev-docker-images` repository will not cause failures in this branch.

This also pins the version of PHP used in the E2E test workflow to PHP 8.0 to avoid deprecated notices related to #54914, which was not fixed until version 6.1.

Follow up to [57198], [57199], [57200].

See #60095.

git-svn-id: https://develop.svn.wordpress.org/branches/5.9@57201 602fd350-edb4-49c9-b593-d223f7449a82