Build/Test Tools: Change the default value of `LOCAL_PHP`.
This changes the default value for `LOCAL_PHP` in the 6.2 branch from `latest` to `8.2-fpm` to reflect the highest version of PHP this branch will support (with beta support).

After this change, future updates to the `latest` container built and published by the `wpdev-docker-images` repository will not cause failures in this branch.

Follow up to [57198], [57199], [57200], [57201], [57202], [57205].

See #60095.

git-svn-id: https://develop.svn.wordpress.org/branches/6.2@57206 602fd350-edb4-49c9-b593-d223f7449a82