Login and Registration: Use correct variable to display a single error message.
When the `wp_login_errors` filter is used to add an error message followed by an instructional message, the former was not displayed and the latter was errouneously displayed twice above the login form.

This commit ensures that the error message is displayed as expected.

Follow-up to [56654].

Reviewed by jorbin.
Merges [57142] to the 6.4 branch.

Props mapumba, SergeyBiryukov.
Fixes #59983.




git-svn-id: https://develop.svn.wordpress.org/branches/6.4@57288 602fd350-edb4-49c9-b593-d223f7449a82