Remove filter for option_autoload_value in option.php
The filter for `option_autoload_value` in the function `wp_determine_option_autoload_value` has been removed. Previously, it was used to determine the maximum size of option value in bytes. This change simplifies the code base and may improve execution speed as the filter processing is no longer required.