HTML API: Cleanup tests and list of void elements.
This patch adds newly supported elements to tests that should have
been updated in recent PRs, but which were merged without that. Those
PRs removed failing tests showing that the elements were unsupported,
but did not add the elements to the list of supported ones.

It also removes some elements from the special-exclusion list of
unsupported IN BODY elements. These did not present in failing tests
because earlier conditions in the switch structure caught the tags
before hitting the default block.

Finally it adds some missing elements to the list of void elements.
These elements are not listed as void in the HTML specification because
they are deprecated. However, they are treated as void for the sake of
HTML serialization and the parsing rules indicate that they behave as
void elements, so it's safe to list them within the HTML API as void.