Ensure void and self-closing elements pop from stack when advancing.
Previously, the logic to pop void and self-closing elements from the
stack of open elements only ran when stepping into the next node in
a document. With the introduction of `next_token()` there appeared a
new way to reprocesses the current token, so this logic would be skipped
when calling `next_token()` _into_ a void or self-closing element,
leaving it on the stack.

In this patch the logic runs whenever the processor is not reprocessing
the current token. A new class constant communicates that `step()`
should treat the current token as if it arrived there itself, that is,
to process it with the normal rules but without advancing the parser.