HTML API: Indicate if a given tag name represents an HTML tag.
The HTML5 living specification defines three kinds of elements:

 - HTML elements.
 - Foreign elements, those which are part of MathML and SVG content.
 - Custom elements.

The HTML API should be able to indicate if a given tag name is possibly
an HTML element or not, so that dependent functions can make appropriate
decisions for self-closing elements.

HTML treats the self-closing flag differently for HTML and foreign
elements. In HTML elements the flag is ignored, but in foreign content
the flag indicates that the tag is self-closing or void. HTML itself
does not contain self-closing elements; only some void elements.

This patch introduces `WP_HTML_Processor::is_html_tag( $tag_name )` to
answer this question.