Add test and implement wp_autoload_values_to_autoload function
This commit adds a new function `wp_autoload_values_to_autoload` in `option.php` to convert autoload values from the options table to string or an array. It also implements its usage in the query for autoloaded options. Additionally, a new unit test for this function is added to `wpAutoloadValuesToAutoload.php`. This provides a more flexible and efficient way of defining which options are to be autoloaded.