HTML API: Reset parser state after seeking to bookmark.
When parser states were introduced in WordPress/wordpress-develop#5683,
nothing in the `seek()` method reset the parser state. This is
problematic because it could leave the parser in the wrong state.

In this patch the parser state is reset so that it get's properly
adjusted on the successive call to `next_token()`.

Follows [57348]

Props @kevin940726 for finding and reporting.