HTML API: Trigger active format reconstruction when reaching text nodes.
When encountering text nodes in an HTML document, the HTML parser needs
to run the active format reconstruction algorithm, even if it doesn't
stop to visit those text nodes. This is because the formats, which might
need reconstructing, will impact the breadcrumbs of all downstream nodes
from the text node.

In this patch, this process is triggered, but the text nodes are then
skipped, since the HTML Processor doesn't currently support visiting
them.

Co-authored-by: Weston Ruter <westonruter@git.wordpress.org>
Co-authored-by: Jon Surrell <jonsurrell@git.wordpress.org>