Move deactivation and persist notices.
Switches from not directly update the active plugins options
to using deactivate_plugins() via "admin_init".

Reverts the performance commit of using a lambda for the admin notice
as the notices need to persist to ensure they are rendered in
wp-admin. The option path persists between page loads, so that
when a user logs in or moves from frontend to backend, the notice
is shown.