Replaced Hard-Coded Test Domains
There were a number of tests that failed when `WP_TESTS_DOMAIN`
has a port. This commit fixes all of those tests so that they will pass
with both bare hostnames and those with ports.

To discover these problems I added a port to my local
`wp-tests-config.php` and ran the test suite.