Site Icon/Customizer: Update strings in customizer to align with `options-general.php`.
This uses the new strings for alt text and site icon description that were introduced in [57713] as a part of #54370.

The strings are translated in PHP and then use `wp.i18n.sprintf` since the alt text strings live in a PHP file even though they are output as part of a template that is used in JavaScript.

Props westonruter, swissspidy, jorbin.
Fixes #60641.


git-svn-id: https://develop.svn.wordpress.org/trunk@57730 602fd350-edb4-49c9-b593-d223f7449a82